requirements.txt
flask
pyserial
flask-socketio
flask-cors
python-dotenv
gevent
torch
huggingface_hub
pillow
transformers